+++ config/tenant.php
@@ -2,20 +2,24 @@
 
 return [
     // Use local tenant configuration files
-    'local' => true,
+    'local' => env('BROKER_LOCAL', true),
 
     // Not found return message
     'not_found' => 'Tenant not found',
 
-    // Header names passed by broker applications
-    'header_host' => 'x-tenant-host',
+    // globally prepend midleware to the specified groups
+    'middleware_groups' => [
+        'api', 'web'
+    ],
+
+    // Headers passed by broker applications
+    'header_alias' => 'x-tenant-alias',
     'header_token' => 'x-tenant-token',
 
     'token' => env('BROKER_API_TOKEN', null),
 
     // Name for auto bind tenant route parameter
     'param' => 'tenant',
-    'domain' => env('BROKER_DOMAIN', 'localhost'),
 
     // Broker connection driver
     'connection' => 'mysql',